refactor: incorrect package refactor to use componentscan QUZ-37
