+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/redis/RedisPublisher.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.infra.game.redis
+package com.grepp.quizy.game.infra.redis
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.data.redis.core.RedisTemplate
@@ -15,4 +15,5 @@ class RedisPublisher(
         val messageJson = objectMapper.writeValueAsString(message)
         redisTemplate.convertAndSend(gameTopic.topic, messageJson)
     }
+
 }
