+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/redis/config/RedisPubSubConfig.kt
@@ -1,17 +1,13 @@
-package com.grepp.quizy.infra.game.redis.config
+package com.grepp.quizy.game.infra.redis.config
 
-import com.grepp.quizy.infra.game.redis.RedisSubscriber
+import com.grepp.quizy.game.infra.redis.RedisSubscriber
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Import
 import org.springframework.data.redis.connection.RedisConnectionFactory
 import org.springframework.data.redis.listener.ChannelTopic
 import org.springframework.data.redis.listener.RedisMessageListenerContainer
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter
 
-@Import(RedisRepositoryConfig::class)
-@ComponentScan(basePackageClasses = [RedisSubscriber::class])
 @Configuration
 class RedisPubSubConfig {
