+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/websocket/CustomHandshakeHandler.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.infra.game.websocket
+package com.grepp.quizy.game.infra.websocket
 
 import org.springframework.http.server.ServerHttpRequest
 import org.springframework.web.socket.WebSocketHandler
