+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/websocket/WebSocketPrincipal.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.infra.game.websocket
+package com.grepp.quizy.game.infra.websocket
 
 import java.security.Principal
